132133/requirements.txt
python-telegram-bot==21.3
pandas==2.1.4
openpyxl==3.11.0
python-dotenv==1.0.0
